add histogram and parallel sort in this commit.

+++ src/common/m_buffer.cpp
@@ -122,8 +122,8 @@ void Block::build(int size, int number) {
 
 int Block::compare_start_end(vector<int> bounds) {
 	int ret;
-	int start=*(int *)((char *)start_+8);
-	int end=*(int *)((char *)start_+tuple_size_*(get_size()-1)+8);
+	unsigned long start=*(unsigned long *)((char *)start_+8);
+	unsigned long end=*(unsigned long *)((char *)start_+tuple_size_*(get_size()-1)+8);
 	for(int i=0; i<bounds.size(); i++) {
 		if(start<bounds[i] && end<bounds[i]) {
 			ret=i;
